changed amount to long

+++ BitcoinBook/TransactionOutput.cs
@@ -4,10 +4,10 @@ namespace BitcoinBook
 {
     public class TransactionOutput
     {
-        public int Amount { get; }
+        public long Amount { get; }
         public ScriptPubKey ScriptPubKey { get; }
 
-        public TransactionOutput(int amount, ScriptPubKey scriptPubKey)
+        public TransactionOutput(long amount, ScriptPubKey scriptPubKey)
         {
             Amount = amount;
             ScriptPubKey = scriptPubKey ?? throw new ArgumentNullException(nameof(scriptPubKey));
